Add files via upload
Write a program in C that reads input CPU bursts from the first line of a text file named CPU_BURST.txt. Validate the input numbers whether the numbers are positive intergers or not. Consider the numbers as CPU bur st.If there are 5 positive integers in the first line of the text file then the program treats that argument as required CPU bust for P1, P2, P3, P4, and P5 process and calculate average waiting time and an average turn around time. Consider used scheduling algorithm as SJF and the same arrival time for all the processes.